add toDouble

+++ app/src/main/java/com/example/project7/Balloons.kt
@@ -25,7 +25,7 @@ class Balloons {
         attempts++
         for (b in bloonsList) {
             // find distance from center
-            var d : Double = Math.sqrt((x - b.getX()).pow(2) +
+            var d : Double = Math.sqrt((x - b.getX()).pow(2).toDouble() +
                     (y - b.getY()).pow(2).toDouble())
             Log.w("Balloons", "distance for balloon with x " + b.getX().toString() + " is " + d.toString() + " with radius " + b.getRadius().toString())
             // check if distance is less than radius
